# swagger.yaml
openapi: 3.0.0
info:
  title: react-bootstrap-api
  contact:
    name: janjic zoran
    url: https://zoran-janjic.netlify.app/
    email: janjiczoran23@gmail.com
  version: '1.0'
servers:
- url: https://react-bootstrap-website-api.herokuapp.com/api/v1
  variables: {}
paths:
...
  /auth/register:
    post:
      tags:
      - Auth
      summary: Register user
      operationId: Registeruser
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/RegisteruserRequest'
              - example:
                  name: Firstname Lastname
                  email: yourEmail@gmail.com
                  password: password
            example:
              name: First Name
              email: yourEmail@gmail.com
              password: password
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security: []
...
components:
  schemas:
...
    RegisteruserRequest:
      title: RegisteruserRequest
      required:
      - name
      - email
      - password
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
      example:
        name: Firstname Lastname
        email: yourEmail@gmail.com
        password: password
...
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
security: []
...